Refacto config key binding et graphics_manager pour l'affichage de texte centré

+++ src/game/car/mod.rs
@@ -9,7 +9,7 @@ pub mod bot_manager;
 pub const PLAYER_CAR_HEIGHT: f32 = 73.0;
 pub const PLAYER_CAR_WIDTH: f32 = 155.0;
 pub const BOT_CAR_WIDTH: f32 = 140.0;
-pub const PLAYER_CAR_X_POSITION: f32 = WINDOW_WIDTH as f32 / 4.0;
+pub const PLAYER_CAR_X_POSITION: f32 = WINDOW_WIDTH / 4.0;
 
 #[derive(PartialEq, Hash, Eq, Clone, Copy)]
 pub enum Way {
@@ -18,6 +18,8 @@ pub enum Way {
     Lower,
 }
 
+/// Trait for all cars, but it is not a entire generic type
+/// to separate the player car and the bot car
 pub trait Car {
     fn get_texture(&self) -> Texture2D;
     fn get_way(&self) -> Way;
